remove dependency cycle
