+++ frontend/util/general_util.js
@@ -34,22 +34,3 @@ export const itemAgeToString = (item) => {
 // takes in blocks
 export const timeDiff = (curr, prev) => curr.timestamp - prev.timestamp;
 
-export const calculateUpdatedRewad = (block, transaction) => {
-  // some validating
-  if (block && block.reward && !transaction) return block.reward;
-  if (!block) return '2';
-  // if (!block.reward) return '2';
-  const costEthe = web3.utils.fromWei(
-    transaction.costOfGasUsed.toString(),
-    'ether',
-  );
-  const bigCost = new Big(costEthe, 10);
-
-  // if for some reason block does not have reward assume its 2
-
-  const blockReward = block.reward ? block.reward : 2;
-  const bigReward = new Big(blockReward, 10).toString();
-
-  const newReward = bigCost.add(bigReward);
-  return newReward.toString();
-}
